# ARROW-11045: [Rust] Fix performance issues of allocator