--- a/rust/arrow/src/array/array_boolean.rs
+++ b/rust/arrow/src/array/array_boolean.rs
@@ -147,7 +147,7 @@ impl From<ArrayDataRef> for BooleanArray {
             1,
             "BooleanArray data should contain a single buffer only (values buffer)"
         );
-        let raw_values = data.buffers()[0].raw_data();
+        let raw_values = data.buffers()[0].ptr();
         assert!(
             memory::is_aligned::<u8>(raw_values, mem::align_of::<bool>()),
             "memory is not aligned"