--- a/rust/arrow/src/array/array_boolean.rs
+++ b/rust/arrow/src/array/array_boolean.rs
@@ -24,7 +24,6 @@ use std::{convert::From, sync::Arc};
 use super::*;
 use super::{array::print_long_array, raw_pointer::RawPtrBox};
 use crate::buffer::{Buffer, MutableBuffer};
-use crate::memory;
 use crate::util::bit_util;
 
 /// Array of bools
@@ -58,7 +57,7 @@ impl BooleanArray {
 
     /// Returns a raw pointer to the values of this array.
     pub fn raw_values(&self) -> *const u8 {
-        unsafe { self.raw_values.get().add(self.data.offset()) }
+        unsafe { self.raw_values.as_ptr().add(self.data.offset()) }
     }
 
     /// Returns a slice for the given offset and length
@@ -87,7 +86,7 @@ impl BooleanArray {
     /// Note this doesn't do any bound checking, for performance reason.
     pub fn value(&self, i: usize) -> bool {
         let offset = i + self.offset();
-        unsafe { bit_util::get_bit_raw(self.raw_values.get() as *const u8, offset) }
+        unsafe { bit_util::get_bit_raw(self.raw_values.as_ptr(), offset) }
     }
 }
 
@@ -119,7 +118,7 @@ impl From<Vec<bool>> for BooleanArray {
     fn from(data: Vec<bool>) -> Self {
         let mut mut_buf = MutableBuffer::new_null(data.len());
         {
-            let mut_slice = mut_buf.data_mut();
+            let mut_slice = mut_buf.as_slice_mut();
             for (i, b) in data.iter().enumerate() {
                 if *b {
                     bit_util::set_bit(mut_slice, i);
@@ -147,14 +146,10 @@ impl From<ArrayDataRef> for BooleanArray {
             1,
             "BooleanArray data should contain a single buffer only (values buffer)"
         );
-        let raw_values = data.buffers()[0].ptr();
-        assert!(
-            memory::is_aligned::<u8>(raw_values, mem::align_of::<bool>()),
-            "memory is not aligned"
-        );
+        let ptr = data.buffers()[0].as_ptr();
         Self {
             data,
-            raw_values: RawPtrBox::new(raw_values as *const u8),
+            raw_values: unsafe { RawPtrBox::new(ptr) },
         }
     }
 }
@@ -185,11 +180,9 @@ impl<Ptr: Borrow<Option<bool>>> FromIterator<Ptr> for BooleanArray {
         let mut null_buf = MutableBuffer::new(num_bytes).with_bitset(num_bytes, false);
         let mut val_buf = MutableBuffer::new(num_bytes).with_bitset(num_bytes, false);
 
-        let data = unsafe {
-            std::slice::from_raw_parts_mut(val_buf.raw_data_mut(), val_buf.capacity())
-        };
+        let data = val_buf.as_slice_mut();
 
-        let null_slice = null_buf.data_mut();
+        let null_slice = null_buf.as_slice_mut();
         iter.enumerate().for_each(|(i, item)| {
             if let Some(a) = item.borrow() {
                 bit_util::set_bit(null_slice, i);