… rust/arrow/Cargo.toml
… rust/arrow/src/array/array_union.rs
… rust/arrow/src/array/builder.rs
… rust/arrow/src/array/data.rs
… rust/arrow/src/array/equal/decimal.rs
… rust/arrow/src/array/equal/fixed_binary.rs
… rust/arrow/src/array/equal/primitive.rs
… rust/arrow/src/array/equal/structure.rs
… rust/arrow/src/array/equal/utils.rs
… rust/arrow/src/array/equal/variable_size.rs
… rust/arrow/src/array/raw_pointer.rs
… rust/arrow/src/array/transform/boolean.rs
… rust/arrow/src/array/transform/fixed_binary.rs
… rust/arrow/src/array/transform/list.rs
… rust/arrow/src/array/transform/mod.rs
… rust/arrow/src/array/transform/primitive.rs
… rust/arrow/src/array/transform/utils.rs
… rust/arrow/src/array/transform/variable_size.rs
… rust/arrow/src/bitmap.rs
… rust/arrow/src/buffer.rs
… rust/arrow/src/bytes.rs
… rust/arrow/src/compute/kernels/arithmetic.rs
… rust/arrow/src/compute/kernels/comparison.rs
… rust/arrow/src/compute/kernels/sort.rs
… rust/arrow/src/compute/kernels/substring.rs
… rust/arrow/src/compute/kernels/take.rs
… rust/arrow/src/compute/util.rs
… rust/arrow/src/datatypes.rs
… rust/arrow/src/ffi.rs
… rust/arrow/src/ipc/writer.rs